Backend/requirements.txt
djangorestframework-simplejwt
django>=5.2
djangorestframework
drf-yasg
psycopg2-binary
python-dotenv
daphne